+++ app/(tabs)/two.tsx
@@ -1,11 +1,12 @@
 import { StyleSheet } from 'react-native';
-
+import  Camera  from '/GIT/OnlyAcademy/components/camera';
 import EditScreenInfo from '@/components/EditScreenInfo';
 import { Text, View } from '@/components/Themed';
 
 export default function TabTwoScreen() {
   return (
     <View style={styles.container}>
+      <Camera/>
       <Text style={styles.title}>Tab Two</Text>
       <View style={styles.separator} lightColor="#eee" darkColor="rgba(255,255,255,0.1)" />
       <EditScreenInfo path="app/(tabs)/two.tsx" />
